Start refactoring Client tests
Former-commit-id: 004be38875436cb465e07a717d8f64120e2c87c3
Former-commit-id: 695bb5c2f3278d2d4128d050b4fd447efd66aaa3
